+++ src/transpiler.ts
@@ -19,29 +19,108 @@ export interface TypeIdentification {
     name: string
 }
 
+type Primitive = 'any' | 'void' | 'undefined' | 'null' | 'never' | 'string' | 'number' | 'boolean' | 'bigint'
+
+/**
+ * This interface defines a module.
+ *
+ * In order to create a new module it is necessary to implement all these functions. Most of them should be quite simple
+ */
 export interface Module<T> {
+    /**
+     * Should create the structure corresponding to a typescript `any` type
+     */
     buildAny: () => T
+    /**
+     * Should create the structure corresponding to a typescript `Date` type
+     */
     buildDate: () => T
-    buildPrimitive: (typeString: string) => T
+    /**
+     * Should create the structure corresponding to a Primitive @see Primitive
+     * @param typeString the name of the primitive
+     */
+    buildPrimitive: (typeString: Primitive) => T
+    /**
+     * Should create an array where each element is of the resolved type
+     * @param resolvedType the type that was resolved for an element in the current array
+     */
     buildArray: (resolvedType: T) => T
+    /**
+     * Should create a tuple like `[number, string, Object]`
+     * @param resolvedTypes Array with each of the members of the tuple
+     */
     buildTuple: (resolvedTypes: T[]) => T
+    /**
+     * Should create the structure corresponding to a typescript Union like `User | Error`
+     * @param resolvedTypes Array with each of the members of the union
+     *                      The order does not matter much here
+     */
     buildUnion: (resolvedTypes: T[]) => T
+    /**
+     * Should create the structure corresponding to a typescript Intersection like `User & UserTeams`
+     * @param resolvedTypes Array with each of the members of the intersection
+     *                      The order matters in TS so make sure that's covered)
+     */
     buildIntersection: (resolvedTypes: T[]) => T
+    /**
+     * Should create the structure corresponding to a typescript Enum like `enum Direction { Up=1, Down, Left, Right }`
+     * @param resolvedTypes Array with each of the members of the enum represented by a `[key, value]`
+     */
     buildEnum: (resolvedTypes: Array<[string, T]>) => T
+    /**
+     * Should create the structure corresponding to a literal value like `5` or `'hello'`
+     * @param literal the literal value
+     */
     buildLiteral: (literal: string | number | boolean | bigint) => T
+    /**
+     * Should create the structure corresponding to an object like `{ foo: 1, goo: Object }`
+     * @param properties an array of the resolved properties of the type
+     * @param type the unique identifier for this type (name + id).
+     *             Please note that the id is only unique within one invocation of the process function.
+     *             This means that calling process twice on the same file does not guarantee that the ids will stay
+     *             the same. These ids are exposed by typescript. And can be used to keep track of recurring type
+     *             definitions.
+     *             We recommend keeping a `Map<name, T>` of all the definitions so they can be used to detect recurrence
+     *             withing the type definitions. See the sample modules for more details on possible implementations.
+     */
     buildObject: (properties: Array<ResolvedProperty<T>>, type: TypeIdentification) => T
+    /**
+     * Should create a reference to a type previously visited
+     * @param type the identification for the type to reference
+     */
     buildReference: (type: TypeIdentification) => T
+    /**
+     * Should create the structure corresponding to a indexable object like `{ [key: K]: V }
+     * @param resolvedType The type V in the example above
+     * @param kind The type K in the example above (can only be string or number)
+     */
     buildIndexableObject: (resolvedType: T, kind: 'string' | 'number') => T
+    /**
+     * Called to end the resolution - in case you have steps that you want to perform when the type is completely
+     * resolved. This is where we usually augment the type definition with the references (definitions)
+     * @param resolvedType The type that was resolved
+     */
     endResolution: (resolvedType: T) => T
+    /**
+     * Returns the resolved type and the definitions map. This might not apply everywhere but is convenient to have
+     * when we want to outsource merging the type and the definitions to the user so that they can merge definitions
+     * from many different nodes
+     * @param resolvedType The type resolved so far
+     */
+    endResolutionWithDefinitions: (resolvedType: T) => { resolvedType: T; definitionsMap: Map<string, T> }
+    /**
+     * Called at the start of the resolution to help you clear any state that you might be keeping internaly
+     */
     startResolution: () => void
 }
 
 export interface Options<T> extends Compiler.Options {
     module: Module<T>
     isProcessable?: (node: ts.Node) => boolean
+    returnDefinitions?: boolean
 }
 
-export function processFiles<T>(options: Options<T>): Array<[string, T]> {
+export function processFiles<T>(options: Options<T>): Array<[string, T, Map<string, T>?]> {
     const { program, checker } = Compiler.createProgram(options)
 
     const nodesOfInterest = program.getSourceFiles().reduce((nodes: ts.Node[], file) => {
@@ -51,8 +130,16 @@ export function processFiles<T>(options: Options<T>): Array<[string, T]> {
 
     return nodesOfInterest.map(typeNode => {
         const fileName = typeNode.getSourceFile().fileName
+
+        options.module.startResolution()
         const type = resolveTypeNode(typeNode, checker, options.module)
-        return [fileName, type] as [string, T]
+
+        if (options.returnDefinitions) {
+            const { resolvedType, definitionsMap } = options.module.endResolutionWithDefinitions(type)
+            return [fileName, resolvedType, definitionsMap]
+        }
+
+        return [fileName, options.module.endResolution(type)]
     })
 }
 
@@ -94,7 +181,7 @@ function resolveTypeNode<T>(startNode: ts.Node, checker: ts.TypeChecker, module:
             return module.buildLiteral(type.value.negative ? `-${type.value.base10Value}` : type.value.base10Value)
         }
         if (Types.isLiteral(type)) return module.buildLiteral(type.value)
-        if (Types.isPrimitive(type)) return module.buildPrimitive(typeString)
+        if (Types.isPrimitive(type)) return module.buildPrimitive(typeString as Primitive)
         if (Types.isArray(type)) return module.buildArray(recursion(type.typeArguments[0]))
         if (Types.isTuple(type)) return module.buildTuple(type.typeArguments.map(recursion))
         if (Types.isEnum(type)) {
@@ -163,7 +250,5 @@ function resolveTypeNode<T>(startNode: ts.Node, checker: ts.TypeChecker, module:
         return 'Not supported' as any
     }
 
-    module.startResolution()
-    const result = recursion(checker.getTypeAtLocation(startNode))
-    return module.endResolution(result)
+    return recursion(checker.getTypeAtLocation(startNode))
 }