+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 - Support for empty objects `{}`. It used to result in a "Not supported" error
 - Support for null and undefined types
+- Support for returning type definitions along with the resolved type
+- Added TSDocs to help with implementing a module
 
 ## [0.1.1] - 2019/02/24